Move list resource
